fix: requests badge showing "00" or wrong count due to string concatenation
PostgreSQL COUNT(*) returns bigint as string. The surgeon action requests
count query was missing ::int cast, causing string concatenation in the
badge (e.g. 0 + "0" = "00", 3 + "6" = "36" > 9 = "9+").

+++ server/storage/surgeonPortal.ts
@@ -226,7 +226,7 @@ export async function getPendingSurgeonActionRequestsCount(
   hospitalId: string,
 ): Promise<number> {
   const [result] = await db
-    .select({ count: sql<number>`COUNT(*)` })
+    .select({ count: sql<number>`COUNT(*)::int` })
     .from(surgeonActionRequests)
     .where(
       and(
